refactor: Controller 수정 및 TableGroupServiceTest 패키지 변경
