+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -28,7 +28,6 @@ public class MenuGroupRestController {
 
         return ResponseEntity.created(uri)
                 .body(createdMenuGroup);
-                .body(created);
     }
 
     @GetMapping("/api/menu-groups")
@@ -37,6 +36,5 @@ public class MenuGroupRestController {
 
         return ResponseEntity.ok()
                 .body(menuGroups);
-                .body(menuGroupService.list());
     }
 }
