+++ src/main/java/kitchenpos/ui/ProductRestController.java
@@ -27,7 +27,6 @@ public class ProductRestController {
         final URI uri = URI.create("/api/products/" + newProduct.getId());
         return ResponseEntity.created(uri)
                 .body(newProduct);
-                .body(created);
     }
 
     @GetMapping("/api/products")
@@ -36,6 +35,5 @@ public class ProductRestController {
 
         return ResponseEntity.ok()
                 .body(products);
-                .body(productService.list());
     }
 }
