+++ src/main/java/kitchenpos/ui/TableGroupRestController.java
@@ -23,7 +23,6 @@ public class TableGroupRestController {
         final URI uri = URI.create("/api/table-groups/" + newTableGroup.getId());
         return ResponseEntity.created(uri)
                 .body(newTableGroup);
-                .body(created);
     }
 
     @DeleteMapping("/api/table-groups/{tableGroupId}")
